+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.141.5 (12/13/2019)
+* Supports Data collection and Data compliance as per GDPR and CCPA.
+* Introduces new API setUserPreferenceForDataCollection() for setting user preferences to opt-out of user data collection.
+* Introduces new API setUserPreferenceForDataDeletion() for setting user preferences to delete previously collected user data.
+
 ## 2.141.4 (08/17/2019)
 * Supports iOS 12.3.
 * Introduces a new API for customers to set CDN Edge Server IP Address.